TAG219 2020/05/30   Core C352     ・C352に対応

+++ mml2vgm/Core/chips/AY8910.cs
@@ -357,8 +357,8 @@ namespace Core
         {
             if (mml.args[0] is string) return;
 
-            int adr = (byte)mml.args[0];
-            byte dat = (byte)mml.args[1];
+            int adr = (byte)(int)mml.args[0];
+            byte dat = (byte)(int)mml.args[1];
 
             parent.OutData(mml, port[0], (byte)adr, (byte)dat);
         }
